Checking out the queryinterface to see what tie95 wants when trying to use 3d hardware, and yup, it wants d3d.

+++ myIDirectDrawSurface.cpp
@@ -290,7 +290,7 @@ HRESULT __stdcall myIDirectDrawSurface::Blt(LPRECT a, LPDIRECTDRAWSURFACE b, LPR
 		  int i, j;
 		  for (i = 0; i < mHeight; i++)
 			  for (j = 0; j < mWidth; j++)
-				  surf[i * mPitch/2 + j] = e->dwFillColor;
+				  surf[i * mPitch/2 + j] = (unsigned short)e->dwFillColor;
 	  }
   }
